+++ Cargo.toml
@@ -30,7 +30,7 @@ harness = false
 name = "io"
 
 [dependencies]
-backoff = { version = "0.4.0", features = ["futures", "tokio"] }
+backon = "0.0.2"
 anyhow = "1.0.56"
 async-compat = "0.2.1"
 async-compression = { version = "0.3.12", features = [
