+++ src/lib.rs
@@ -89,8 +89,8 @@ pub use io::BytesStream;
 pub use io::BytesWrite;
 pub use io::BytesWriter;
 
-mod layer;
-pub use layer::Layer;
+mod layers;
+pub use layers::Layer;
 
 mod operator;
 pub use operator::Operator;
